pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codechat"
version = "1.0.0"
description = "Chat with Your Codebase - RAG system for GitHub repositories"
authors = [{name = "Chat Git Repo", email = "contact@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitpython==3.1.43",
    "langchain==0.2.1", 
    "llama-index==0.11.39",
    "sentence-transformers==2.7.0",
    "faiss-cpu==1.8.0",
    "transformers==4.42.0",
    "accelerate==0.29.3",
    "huggingface-hub==0.23.3",
    "bitsandbytes==0.43.1",
    "pydantic==2.7.3",
    "typer[all]==0.12.3",
    "tqdm==4.66.4"
]

[project.optional-dependencies]
dev = [
    "black==24.4.2",
    "isort==5.13.2", 
    "flake8==7.0.0",
    "mypy==1.10.0",
    "pytest==8.2.2",
    "pytest-cov==5.0.0",
    "coverage==7.5.3"
]

[project.scripts]
codechat = "codechat.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["codechat*"]

[tool.black]
line-length = 120
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=codechat",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=90"
]
markers = [
    "integration: marks tests as integration tests (may be slow)",
    "gpu: marks tests that require GPU"
]
